embed/requirements.txt
python-dotenv
joblib
gensim
numpy
pandas
sklearn
matplotlib